@@ -7,10 +7,10 @@ import (
 )
 
 func TestCircular1(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushBack(1)
 	assert.Assert(t, ok == true)
@@ -28,14 +28,14 @@ func TestCircular1(t *testing.T) {
 	assert.Assert(t, value == 1 && ok == true)
 
 	value, ok = c.PopBack()
-	assert.Assert(t, value == nil && ok == false)
+	assert.Assert(t, value == 0 && ok == false)
 }
 
 func TestCircular2(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushFront(1)
 	assert.Assert(t, ok == true)
@@ -53,14 +53,14 @@ func TestCircular2(t *testing.T) {
 	assert.Assert(t, value == 1 && ok == true)
 
 	value, ok = c.PopFront()
-	assert.Assert(t, value == nil && ok == false)
+	assert.Assert(t, value == 0 && ok == false)
 }
 
 func TestCircular3(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushBack(1)
 	assert.Assert(t, ok == true)
@@ -78,14 +78,14 @@ func TestCircular3(t *testing.T) {
 	assert.Assert(t, value == 2 && ok == true)
 
 	value, ok = c.PopFront()
-	assert.Assert(t, value == nil && ok == false)
+	assert.Assert(t, value == 0 && ok == false)
 }
 
 func TestCircular4(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushFront(1)
 	assert.Assert(t, ok == true)
@@ -103,14 +103,14 @@ func TestCircular4(t *testing.T) {
 	assert.Assert(t, value == 2 && ok == true)
 
 	value, ok = c.PopBack()
-	assert.Assert(t, value == nil && ok == false)
+	assert.Assert(t, value == 0 && ok == false)
 }
 
 func TestCircular5(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushBack(1)
 	assert.Assert(t, ok == true)
@@ -128,14 +128,14 @@ func TestCircular5(t *testing.T) {
 	assert.Assert(t, value == 2 && ok == true)
 
 	value, ok = c.PopBack()
-	assert.Assert(t, value == nil && ok == false)
+	assert.Assert(t, value == 0 && ok == false)
 }
 
 func TestCircular6(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushFront(1)
 	assert.Assert(t, ok == true)
@@ -153,14 +153,14 @@ func TestCircular6(t *testing.T) {
 	assert.Assert(t, value == 2 && ok == true)
 
 	value, ok = c.PopFront()
-	assert.Assert(t, value == nil && ok == false)
+	assert.Assert(t, value == 0 && ok == false)
 }
 
 func TestCircular7(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushFront(1)
 	assert.Assert(t, ok == true)
@@ -170,10 +170,10 @@ func TestCircular7(t *testing.T) {
 }
 
 func TestCircular8(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushBack(1)
 	assert.Assert(t, ok == true)
@@ -183,14 +183,14 @@ func TestCircular8(t *testing.T) {
 }
 
 func TestCircular9(t *testing.T) {
-	c := New(3)
+	c := New[int](3)
 
 	c.PushBack(1)
 	c.PushBack(2)
 	c.PushBack(3)
 	i := 1
 	c.RangeFront(
-		func(v interface{}) bool {
+		func(v int) bool {
 			assert.Assert(t, v == i)
 			i++
 			return true
@@ -198,7 +198,7 @@ func TestCircular9(t *testing.T) {
 	)
 	i = 3
 	c.RangeBack(
-		func(v interface{}) bool {
+		func(v int) bool {
 			assert.Assert(t, v == i)
 			i--
 			return true
@@ -207,14 +207,14 @@ func TestCircular9(t *testing.T) {
 }
 
 func TestCircular10(t *testing.T) {
-	c := New(3)
+	c := New[int](3)
 
 	c.PushFront(3)
 	c.PushFront(2)
 	c.PushFront(1)
 	i := 1
 	c.RangeFront(
-		func(v interface{}) bool {
+		func(v int) bool {
 			assert.Assert(t, v == i)
 			i++
 			return true
@@ -222,7 +222,7 @@ func TestCircular10(t *testing.T) {
 	)
 	i = 3
 	c.RangeBack(
-		func(v interface{}) bool {
+		func(v int) bool {
 			assert.Assert(t, v == i)
 			i--
 			return true
@@ -231,7 +231,7 @@ func TestCircular10(t *testing.T) {
 }
 
 func TestCircular11(t *testing.T) {
-	c := New(3)
+	c := New[int](3)
 
 	c.PushBack(4)
 	c.PushFront(2)
@@ -240,7 +240,7 @@ func TestCircular11(t *testing.T) {
 	c.PushBack(3)
 	i := 1
 	c.RangeFront(
-		func(v interface{}) bool {
+		func(v int) bool {
 			assert.Assert(t, v == i)
 			i++
 			return true
@@ -248,7 +248,7 @@ func TestCircular11(t *testing.T) {
 	)
 	i = 3
 	c.RangeBack(
-		func(v interface{}) bool {
+		func(v int) bool {
 			assert.Assert(t, v == i)
 			i--
 			return true
@@ -257,10 +257,10 @@ func TestCircular11(t *testing.T) {
 }
 
 func TestCircular12(t *testing.T) {
-	var value interface{}
+	var value int
 	var ok bool
 
-	c := New(2)
+	c := New[int](2)
 
 	ok = c.PushBack(1)
 	assert.Assert(t, ok == true)
@@ -292,7 +292,7 @@ func TestCircular12(t *testing.T) {
 
 func BenchmarkCircular1(b *testing.B) {
 	b.ReportAllocs()
-	c := New(b.N)
+	c := New[string](b.N)
 	for i := 0; i < b.N; i++ {
 		c.PushBack("lalala")
 	}
